# Optimize requests