--- a/client/src/index.css
+++ b/client/src/index.css
@@ -282,6 +282,7 @@ form.login {
   font-size: 30px;
   padding-left: 20px;
   color: #616161;
+  padding-right: 20px;
 }
 .subheader {
   box-shadow: