--- a/client/src/queries/Lists.js
+++ b/client/src/queries/Lists.js
@@ -1,13 +1,11 @@
 import { gql } from "@apollo/client";
 
 export default gql`
-  {
-    user {
-      id
-    }
+  query GetLists {
     lists {
       id
       name
+      user
     }
   }
 `;