--- a/server/helpers/index.js
+++ b/server/helpers/index.js
@@ -7,3 +7,6 @@ export {
   filterOutDuplicateItems,
   getChildMediaIds,
 } from "./media.js";
+
+export const isProduction = () => process.env.NODE_ENV === "production";
+export const isTest = () => process.env.NODE_ENV === "test";