--- a/server/schema/types/media_type.js
+++ b/server/schema/types/media_type.js
@@ -7,20 +7,13 @@ import {
 } from "graphql";
 import GraphQLDate from "graphql-date";
 
-import { MediaTypeEnum } from "../../helpers/index.js";
-
 const MediaType = new GraphQLObjectType({
   name: "MediaType",
   fields: () => ({
     id: { type: GraphQLID }, // this is the MLab id
     title: { type: GraphQLString },
     release_date: { type: GraphQLDate },
-    media_type: {
-      type: GraphQLString,
-      resolve(parentValue) {
-        return MediaTypeEnum[parentValue.media_type];
-      },
-    },
+    media_type: { type: GraphQLString },
     poster_path: { type: GraphQLString },
     media_id: { type: GraphQLString }, // this is the TMDB id
     number: { type: GraphQLInt },