This pull requests significantly improves the time for queries and page load. 

| RootQueryType resolver | Before | After |
|-----|----|----|
lists | ~230m | ~125m
media | ~200ms | ~175m

Key changes were the removal of loops within the request, adding indexing, and transporting a minimum of data from the DB.

In addition, previously multiple lists requests were made on each page load. Now the lists request only occurs when the dropdown is first opened, meaning that the header load isn't blocked by the long lists request.